:recycle: 출석 도메인 변경
