+++ .editorconfig
@@ -12,6 +12,7 @@ trim_trailing_whitespace = true
 [*.{kt,kts}]
 ktlint_code_style = ktlint_official
 ktlint_standard_no-consecutive-blank-lines-in-method-declaration = disabled
+ktlint_function_signature_body_expression_wrapping=multiline
 ktlint_standard_no-blank-line-in-list = disabled
 ij_kotlin_name_count_to_use_star_import = 2147483647
 ij_kotlin_name_count_to_use_star_import_for_members = 2147483647
